Added field name translation support (also known as Field Naming Policy).
This CL allows client libraries to specify a different JSON field name output than the actual Java class field name.  Currently, we are only exposing three types of field translations:
1) Default - Java Field Name is used for JSON field name
2) Upper Camel Case - The first letter is changed to upper case and the remaining Java Field name is unchanged
3) Lower Case With Underscore - A Java field name that is camel cased is separated into words using an underscore (_) and then changed to lower case.

The field naming policy is implemented in a generic way that we can mix and match these "strategies"; however, we are not exposing any of this infrastructure except through the "FieldNamingPolicy" enum.  This enum hides the infrastructure from the client so that it does not leak into the public API in its premature state.

git-svn-id: http://google-gson.googlecode.com/svn/trunk@100 2534bb62-2c4b-0410-85e8-b5006b95c4ae